requirements.txt
# Django Backend Requirements
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-decouple==3.8

# Google AI
google-generativeai==0.3.2

# Streamlit (for support system)
streamlit==1.28.1

# Security
cryptography==41.0.7

# Development
python-dotenv==1.0.0
